feat: 지역 검색 무한 스크롤 기능 추가 및 로딩 화면 모달로 수정
- AreaSearch로 컴포넌트 분리
