+++ front-end/legeno-around-here/src/components/pages/ranking/RankingTopBar.js
@@ -14,10 +14,10 @@ import TextField from '@material-ui/core/TextField';
 import Button from '@material-ui/core/Button';
 import SearchIcon from '@material-ui/icons/Search';
 import List from '@material-ui/core/List';
-import Loading from '../../Loading'
-import { getAccessTokenFromCookie } from '../../../util/TokenUtils'
-import { findAllAreas } from '../../api/API'
-import AreaItem from '../../AreaItem'
+import Loading from '../../Loading';
+import { getAccessTokenFromCookie } from '../../../util/TokenUtils';
+import { findAreasFromPage } from '../../api/API';
+import AreaItem from '../../AreaItem';
 
 const useStyles = makeStyles((theme) => ({
   grow: {
@@ -66,7 +66,7 @@ export default function PrimarySearchAppBar() {
   const loadAreas = async () => {
     const accessToken = getAccessTokenFromCookie();
     setLoading(true);
-    const allAreas = await findAllAreas(page, accessToken, areaKeyword);
+    const allAreas = await findAreasFromPage(page, accessToken, areaKeyword);
     if (allAreas.length === 0) {
       alert('검색 결과가 없습니다! 다시 검색해주세요!');
     }
@@ -93,23 +93,23 @@ export default function PrimarySearchAppBar() {
   return (
     <>
       {loading && <Loading />}
-      <AppBar position="sticky">
+      <AppBar position='sticky'>
         <Toolbar>
           <IconButton
-            edge="start"
+            edge='start'
             className={classes.menuButton}
-            color="inherit"
-            aria-label="open drawer"
+            color='inherit'
+            aria-label='open drawer'
             onClick={handleOpen}
           >
             <ExpandMoreIcon />
-            <Typography className={classes.title} variant="h6" noWrap>
+            <Typography className={classes.title} variant='h6' noWrap>
               {mainArea} 캡짱은 누구?
             </Typography>
           </IconButton>
           <Modal
-            aria-labelledby="transition-modal-title"
-            aria-describedby="transition-modal-description"
+            aria-labelledby='transition-modal-title'
+            aria-describedby='transition-modal-description'
             className={classes.modal}
             open={open}
             onClose={handleClose}
@@ -121,19 +121,19 @@ export default function PrimarySearchAppBar() {
           >
             <Fade in={open}>
               <div className={classes.paper}>
-                <h2 id="transition-modal-title">지역을 검색해주세요!</h2>
+                <h2 id='transition-modal-title'>지역을 검색해주세요!</h2>
                 <TextField
-                  id="outlined-search"
-                  label="Search field"
-                  type="search"
-                  variant="outlined"
+                  id='outlined-search'
+                  label='Search field'
+                  type='search'
+                  variant='outlined'
                   onChange={(event) => getInputArea(event)}
                 />
                 <Button>
                   <SearchIcon onClick={() => findAllArea()} />
                 </Button>
                 {areas.length > 0 && (
-                  <List component="nav" className={classes.list}>
+                  <List component='nav' className={classes.list}>
                     {areas.map((area) => (
                       <AreaItem key={area.id} area={area} />
                     ))}
@@ -145,13 +145,13 @@ export default function PrimarySearchAppBar() {
           <div className={classes.grow} />
           <div className={classes.sectionDesktop}>
             <IconButton
-              aria-label="show 17 new notifications"
-              color="inherit"
+              aria-label='show 17 new notifications'
+              color='inherit'
               onClick={() => {
                 alert('아직 알람기능이 완성되지 않았습니다!');
               }}
             >
-              <Badge badgeContent={0} color="secondary">
+              <Badge badgeContent={0} color='secondary'>
                 <NotificationsIcon />
               </Badge>
             </IconButton>
